[EDT2] : correctif sur un token oublié + préparatifs pour refactorisation de la page d'affichage des edt + préparatifs pour la génération auto des périodes de vacances
git-svn-id: http://svn.sylogix.org/gepiphp/trunk@7231 e073496e-d129-0410-8f7b-841bfd467322

+++ edt_organisation/lib/actions/ajaxrequest.php
@@ -33,18 +33,20 @@ class ajaxrequestAction extends Action {
     {
 		$content = "";
 		if ($request->getParam('asker')) {
-			check_token();
-			if ($request->getParam('asker') == "calendrier") {
-				$this->insertPeriod($content, $request);
-			}
-			else if ($request->getParam('asker') == "edit_period") {
-				$this->editPeriod($content, $request);
-			}
-			else if ($request->getParam('asker') == "delete_period") {
-				$this->deletePeriod($content, $request);
-			}
-			else if ($request->getParam('asker') == "validate_period") {
-				$this->validatePeriod($content, $request);
+			if ($_SESSION['statut'] == "administrateur") {
+				check_token();
+				if ($request->getParam('asker') == "calendrier") {
+					$this->insertPeriod($content, $request);
+				}
+				else if ($request->getParam('asker') == "edit_period") {
+					$this->editPeriod($content, $request);
+				}
+				else if ($request->getParam('asker') == "delete_period") {
+					$this->deletePeriod($content, $request);
+				}
+				else if ($request->getParam('asker') == "validate_period") {
+					$this->validatePeriod($content, $request);
+				}
 			}
 		}
 		$response->addVar('content', $content);
